backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==0.19.0
openai==1.1.1
httpx<0.25.0
gunicorn==20.1.0 